set basic as default ui in facefusion.ini
