+++ facefusion.ini
@@ -83,7 +83,7 @@ lip_syncer_model =
 
 [uis]
 open_browser =
-ui_layouts =
+ui_layouts = basic
 ui_workflow =
 
 [execution]
