+++ facefusion/uis/layouts/basic.py
@@ -91,5 +91,5 @@ def listen() -> None:
 
 
 def run(ui : gradio.Blocks) -> None:
-	# ui.launch(favicon_path = 'facefusion.ico', inbrowser = state_manager.get_item('open_browser'))
+	#ui.launch(favicon_path = 'facefusion.ico', inbrowser = state_manager.get_item('open_browser'))
 	ui.launch(inbrowser = state_manager.get_item('open_browser'))
