+++ facefusion/uis/layouts/default.py
@@ -5,91 +5,48 @@ from facefusion.uis.components import about, age_modifier_options, common_option
 
 
 def pre_check() -> bool:
-	state_manager.set_item('advanced_user', False)
-	#age_modifier_options.render() # avoid Key Error at launch
+	state_manager.set_item('advanced_user', True)
 	return True
 
 
-def toggle_source_visibility(selected_processors):
-    # check if one processor need the source window
-    requires_source = any(
-        processor in processors.processors_requiring_source() for processor in selected_processors
-    )
-    return gradio.update(visible=requires_source)
-
 def render() -> gradio.Blocks:
 	with gradio.Blocks() as layout:
-		# Advanced_user visibility functionnalies and blocks
-		def toggle_advanced_user(current_state):
-			state_manager.set_item('advanced_user', current_state)
-			advanced_user = state_manager.get_item('advanced_user') == True
-			return [
-				gradio.update(visible=True), # basic_block1
-				gradio.update(visible=advanced_user), # advanced_block1
-				gradio.update(visible=True), # basic_block2
-				gradio.update(visible=advanced_user), # advanced_block2
-				gradio.update(visible=advanced_user), # advanced_block3
-				gradio.update(visible=advanced_user), # age_modifier_stride slider
-				gradio.update(visible=advanced_user), # face_swapper_model_dropdown
-				gradio.update(visible=advanced_user), # log_level dropdown
-				gradio.update(visible=advanced_user), # face_mask_blur slider
-				gradio.update(visible=advanced_user), # output output_path
-			]
-		
 		with gradio.Row():
-			# Column 1
-			with gradio.Column(scale=4):
+			with gradio.Column(scale = 4):
 				with gradio.Blocks():
 					about.render()
-				
-				# Switch for Advanced User
-				with gradio.Blocks():
-					advanced_user_switch = gradio.Checkbox(
-					label="Advanced User",
-					value=state_manager.get_item('advanced_user') == True,
-					interactive=True
-					)
-
 				with gradio.Blocks():
 					processors.render()
-
-				with gradio.Group(visible=True) as basic_block1:
-					with gradio.Blocks():
-						age_modifier_options.render()
-					with gradio.Blocks():
-						expression_restorer_options.render()
-					with gradio.Blocks():
-						face_debugger_options.render()
-					with gradio.Blocks():
-						face_editor_options.render()
-					with gradio.Blocks():
-						face_enhancer_options.render()
-					with gradio.Blocks():
-						face_swapper_options.render()
-					with gradio.Blocks():
-						frame_colorizer_options.render()
-					with gradio.Blocks():
-						frame_enhancer_options.render()
-					with gradio.Blocks():
-						lip_syncer_options.render()
-
-				with gradio.Group(visible=state_manager.get_item('advanced_user')) as advanced_block1:
-					with gradio.Blocks():
-						execution.render()
-						execution_thread_count.render()
-						execution_queue_count.render()
-					with gradio.Blocks():
-						memory.render()
-
-				with gradio.Group(visible=True) as basic_block2:
-					with gradio.Blocks():
-						temp_frame.render()
-					with gradio.Blocks():
-						output_options.render()
-
-			# Column 2
+				with gradio.Blocks():
+					age_modifier_options.render()
+				with gradio.Blocks():
+					expression_restorer_options.render()
+				with gradio.Blocks():
+					face_debugger_options.render()
+				with gradio.Blocks():
+					face_editor_options.render()
+				with gradio.Blocks():
+					face_enhancer_options.render()
+				with gradio.Blocks():
+					face_swapper_options.render()
+				with gradio.Blocks():
+					frame_colorizer_options.render()
+				with gradio.Blocks():
+					frame_enhancer_options.render()
+				with gradio.Blocks():
+					lip_syncer_options.render()
+				with gradio.Blocks():
+					execution.render()
+					execution_thread_count.render()
+					execution_queue_count.render()
+				with gradio.Blocks():
+					memory.render()
+				with gradio.Blocks():
+					temp_frame.render()
+				with gradio.Blocks():
+					output_options.render()
 			with gradio.Column(scale = 4):
-				with gradio.Group(visible=True) as source_group:
+				with gradio.Blocks():
 					source.render()
 				with gradio.Blocks():
 					target.render()
@@ -97,17 +54,11 @@ def render() -> gradio.Blocks:
 					output.render()
 				with gradio.Blocks():
 					terminal.render()
-
-				with gradio.Group(visible=state_manager.get_item('advanced_user')) as advanced_block2:
-					with gradio.Blocks():
-						ui_workflow.render()
-				
 				with gradio.Blocks():
+					ui_workflow.render()
 					instant_runner.render()
 					job_runner.render()
 					job_manager.render()
-
-			# Column 3
 			with gradio.Column(scale = 7):
 				with gradio.Blocks():
 					preview.render()
@@ -117,34 +68,12 @@ def render() -> gradio.Blocks:
 					face_selector.render()
 				with gradio.Blocks():
 					face_masker.render()
-
-				with gradio.Group(visible=state_manager.get_item('advanced_user')) as advanced_block3:
-					with gradio.Blocks():
-						face_detector.render()
-					with gradio.Blocks():
-						face_landmarker.render()
-					with gradio.Blocks():
-						common_options.render()
-
-		# Groups update
-		advanced_user_switch.change(
-			toggle_advanced_user,
-			inputs=[advanced_user_switch],
-			outputs=[basic_block1, advanced_block1, basic_block2, advanced_block2, advanced_block3, 
-			age_modifier_options.AGE_MODIFIER_STRIDE_SLIDER,
-			face_swapper_options.FACE_SWAPPER_MODEL_DROPDOWN,
-			terminal.LOG_LEVEL_DROPDOWN,
-			face_masker.FACE_MASK_BLUR_SLIDER,
-			output.OUTPUT_PATH_TEXTBOX
-			]
-		)
-
-		# Connect Processors to Source
-		processors.PROCESSORS_CHECKBOX_GROUP.change(
-			toggle_source_visibility,
-			inputs=[processors.PROCESSORS_CHECKBOX_GROUP],
-			outputs=[source_group]
-		)
+				with gradio.Blocks():
+					face_detector.render()
+				with gradio.Blocks():
+					face_landmarker.render()
+				with gradio.Blocks():
+					common_options.render()
 	return layout
 
 
@@ -159,11 +88,16 @@ def listen() -> None:
 	frame_colorizer_options.listen()
 	frame_enhancer_options.listen()
 	lip_syncer_options.listen()
+	execution.listen()
+	execution_thread_count.listen()
+	execution_queue_count.listen()
+	memory.listen()
 	temp_frame.listen()
 	output_options.listen()
 	source.listen()
 	target.listen()
 	output.listen()
+	instant_runner.listen()
 	job_runner.listen()
 	job_manager.listen()
 	terminal.listen()
@@ -171,15 +105,9 @@ def listen() -> None:
 	trim_frame.listen()
 	face_selector.listen()
 	face_masker.listen()
-	execution.listen()
-	execution_thread_count.listen() 
-	execution_queue_count.listen()
-	memory.listen()
-	instant_runner.listen()
 	face_detector.listen()
 	face_landmarker.listen()
 	common_options.listen()
-
 
 
 def run(ui : gradio.Blocks) -> None:
